+++ test/libsolidity/syntaxTests/parsing/address_payable.sol
@@ -1,0 +1,7 @@
+contract C {
+    address payable a;
+    function f(address payable b) public pure returns (address payable c) {
+        address payable d = b;
+        return d;
+    }
+}