+++ test/libsolidity/syntaxTests/types/address/address_to_contract_payable_fallback.sol
@@ -1,0 +1,8 @@
+contract C {
+  function f() public pure returns (C c) {
+    c = C(address(2));
+  }
+  function() external payable {
+  }
+}
+// ----