+++ test/libsolidity/syntaxTests/types/address/address_tuple_fail.sol
@@ -1,0 +1,8 @@
+contract C {
+    function f() public view returns (address payable a, address b) {
+        (address c, address payable d) = (address(this), address(0));
+        (a,b) = (c,d);
+    }
+}
+// ----
+// TypeError: (169-174): Type tuple(address,address payable) is not implicitly convertible to expected type tuple(address payable,address).