+++ test/libsolidity/syntaxTests/types/address/literal_to_address.sol
@@ -1,0 +1,8 @@
+contract C {
+    function f() public pure {
+        address a = address(0);
+        a = address(1);
+        address b = 0x0123456789012345678901234567890123456789;
+        b = 0x9876543210987654321098765432109876543210;
+    }
+}